+++ test/Logger/LoggerFactoryTest.php
@@ -5,15 +5,16 @@ declare(strict_types=1);
 namespace ShlinkioTest\Shlink\Common\Logger;
 
 use Monolog\Formatter\LineFormatter;
+use Monolog\Handler\AbstractHandler;
+use Monolog\Handler\FormattableHandlerInterface;
+use Monolog\Handler\ProcessableHandlerInterface;
 use Monolog\Handler\RotatingFileHandler;
 use Monolog\Handler\StreamHandler;
 use Monolog\Level;
 use Monolog\Logger;
 use PHPUnit\Framework\Assert;
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use Prophecy\Argument;
-use Prophecy\PhpUnit\ProphecyTrait;
-use Prophecy\Prophecy\ObjectProphecy;
 use Psr\Container\ContainerInterface;
 use Shlinkio\Shlink\Common\Logger\Exception\InvalidLoggerException;
 use Shlinkio\Shlink\Common\Logger\LoggerFactory;
@@ -23,28 +24,25 @@ use function Functional\id;
 
 class LoggerFactoryTest extends TestCase
 {
-    use ProphecyTrait;
-
-    private ObjectProphecy $container;
+    private MockObject & ContainerInterface $container;
 
     public function setUp(): void
     {
-        $this->container = $this->prophesize(ContainerInterface::class);
+        $this->container = $this->createMock(ContainerInterface::class);
     }
 
     /** @test */
     public function anExceptionIsThrownWhenRequestedLoggerDoesNotHaveConfig(): void
     {
-        $getConfig = $this->container->get('config')->willReturn(['logger' => []]);
+        $this->container->expects($this->once())->method('get')->with('config')->willReturn(['logger' => []]);
 
-        $getConfig->shouldBeCalledOnce();
         $this->expectException(InvalidLoggerException::class);
         $this->expectExceptionMessage(
             'Provided logger with name "foo" is not valid. Make sure to provide a value defined under the "logger" '
             . 'config key.',
         );
 
-        LoggerFactory::foo($this->container->reveal());
+        LoggerFactory::foo($this->container); // @phpstan-ignore-line
     }
 
     /**
@@ -53,15 +51,14 @@ class LoggerFactoryTest extends TestCase
      */
     public function anExceptionIsThrownWhenConfiguredTypeIsInvalid(array $config): void
     {
-        $getConfig = $this->container->get('config')->willReturn([
+        $this->container->expects($this->once())->method('get')->with('config')->willReturn([
             'logger' => ['foo' => $config],
         ]);
 
-        $getConfig->shouldBeCalledOnce();
         $this->expectException(InvalidLoggerException::class);
         $this->expectExceptionMessage('Expected one of ["file", "stream"]');
 
-        LoggerFactory::foo($this->container->reveal());
+        LoggerFactory::foo($this->container); // @phpstan-ignore-line
     }
 
     public function provideConfigWithInvalidType(): iterable
@@ -71,23 +68,24 @@ class LoggerFactoryTest extends TestCase
     }
 
     /**
+     * @param class-string<ProcessableHandlerInterface> $expectedHandler
      * @test
      * @dataProvider provideTypes
      */
     public function expectedHandlerIsCreated(array $config, string $expectedHandler, callable $assertConfig): void
     {
-        $getConfig = $this->container->get('config')->willReturn([
+        $this->container->expects($this->once())->method('get')->with('config')->willReturn([
             'logger' => ['foo' => $config],
         ]);
 
         /** @var Logger $logger */
-        $logger = LoggerFactory::foo($this->container->reveal());
+        $logger = LoggerFactory::foo($this->container); // @phpstan-ignore-line
         $handlers = $logger->getHandlers();
 
         self::assertCount(1, $handlers);
         self::assertInstanceOf($expectedHandler, $handlers[0]);
+        self::assertInstanceOf(FormattableHandlerInterface::class, $handlers[0]);
         self::assertInstanceOf(LineFormatter::class, $handlers[0]->getFormatter());
-        $getConfig->shouldHaveBeenCalledOnce();
         $assertConfig($handlers[0]);
     }
 
@@ -97,16 +95,22 @@ class LoggerFactoryTest extends TestCase
             ['type' => LoggerType::FILE->value],
             RotatingFileHandler::class,
             static function (RotatingFileHandler $handler): void {
+                $url = $handler->getUrl();
+
                 Assert::assertTrue($handler->getBubble());
-                Assert::assertStringContainsString('data/log/shlink_log', $handler->getUrl());
+                Assert::assertNotNull($url);
+                Assert::assertStringContainsString('data/log/shlink_log', $url);
             },
         ];
         yield [
             ['type' => LoggerType::FILE->value, 'destination' => 'foobar'],
             RotatingFileHandler::class,
             static function (RotatingFileHandler $handler): void {
+                $url = $handler->getUrl();
+
                 Assert::assertTrue($handler->getBubble());
-                Assert::assertStringContainsString('foobar', $handler->getUrl());
+                Assert::assertNotNull($url);
+                Assert::assertStringContainsString('foobar', $url);
             },
         ];
         yield [
@@ -131,20 +135,17 @@ class LoggerFactoryTest extends TestCase
      */
     public function extraProcessorsAreAdded(array $config, int $expectedAmountOfProcessors): void
     {
-        $getConfig = $this->container->get('config')->willReturn(['logger' => [
-            'foo' => ['type' => LoggerType::STREAM->value, ...$config],
-        ]]);
-        $getProcessor = $this->container->get(
-            Argument::that(static fn (string $arg) => $arg !== 'config'),
-        )->willReturn(id(...));
+        $this->container->expects($this->exactly($expectedAmountOfProcessors + 1))->method('get')->willReturnCallback(
+            fn (string $serviceName) => $serviceName !== 'config' ? id(...) : ['logger' => [
+                'foo' => ['type' => LoggerType::STREAM->value, ...$config],
+            ]],
+        );
 
         /** @var Logger $logger */
-        $logger = LoggerFactory::foo($this->container->reveal());
+        $logger = LoggerFactory::foo($this->container); // @phpstan-ignore-line
         $processors = $logger->getProcessors();
 
         self::assertCount($expectedAmountOfProcessors + 2, $processors);
-        $getConfig->shouldHaveBeenCalledOnce();
-        $getProcessor->shouldHaveBeenCalledTimes($expectedAmountOfProcessors);
     }
 
     public function provideExtraProcessors(): iterable
@@ -161,14 +162,17 @@ class LoggerFactoryTest extends TestCase
      */
     public function expectedLevelIsSetBasedOnConfig(array $config, Level $expectedLevel): void
     {
-        $this->container->get('config')->willReturn(['logger' => [
+        $this->container->method('get')->with('config')->willReturn(['logger' => [
             'bar' => ['type' => LoggerType::STREAM->value, ...$config],
         ]]);
 
         /** @var Logger $logger */
-        $logger = LoggerFactory::bar($this->container->reveal());
+        $logger = LoggerFactory::bar($this->container); // @phpstan-ignore-line
+        $handlers = $logger->getHandlers();
 
-        self::assertEquals($expectedLevel, $logger->getHandlers()[0]->getLevel());
+        self::assertNotEmpty($handlers);
+        self::assertInstanceOf(AbstractHandler::class, $handlers[0]);
+        self::assertEquals($expectedLevel, $handlers[0]->getLevel());
     }
 
     public function provideLevelConfig(): iterable