+++ test/Mercure/HubFactoryTest.php
@@ -4,9 +4,8 @@ declare(strict_types=1);
 
 namespace ShlinkioTest\Shlink\Common\Mercure;
 
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use Prophecy\PhpUnit\ProphecyTrait;
-use Prophecy\Prophecy\ObjectProphecy;
 use Psr\Container\ContainerInterface;
 use ReflectionObject;
 use Shlinkio\Shlink\Common\Exception\MercureException;
@@ -16,16 +15,12 @@ use Symfony\Component\Mercure\Jwt\TokenProviderInterface;
 
 class HubFactoryTest extends TestCase
 {
-    use ProphecyTrait;
-
     private HubFactory $factory;
-    private ObjectProphecy $container;
-    private ObjectProphecy $tokenProvider;
+    private MockObject & ContainerInterface $container;
 
     public function setUp(): void
     {
-        $this->container = $this->prophesize(ContainerInterface::class);
-        $this->tokenProvider = $this->prophesize(TokenProviderInterface::class);
+        $this->container = $this->createMock(ContainerInterface::class);
         $this->factory = new HubFactory();
     }
 
@@ -35,17 +30,14 @@ class HubFactoryTest extends TestCase
      */
     public function throwsExceptionWhenNoHubUrlIsConfigured(array $config): void
     {
-        $getConfig = $this->container->get('config')->willReturn($config);
-        $getJwtProvider = $this->container->get(LcobucciJwtProvider::class)->willReturn($this->tokenProvider->reveal());
+        $this->container->expects($this->once())->method('get')->with('config')->willReturn($config);
 
         $this->expectException(MercureException::class);
         $this->expectExceptionMessage(
             'You have to provide mercure hub URL under mercure.internal_hub_url or mercure.public_hub_url',
         );
-        $getConfig->shouldBeCalledOnce();
-        $getJwtProvider->shouldNotBeCalled();
 
-        ($this->factory)($this->container->reveal());
+        ($this->factory)($this->container);
     }
 
     public function provideInvalidConfigs(): iterable
@@ -70,18 +62,18 @@ class HubFactoryTest extends TestCase
      */
     public function returnsExpectedObjectIfProperConfigIsFound(array $config, string $expectedHubUrl): void
     {
-        $getConfig = $this->container->get('config')->willReturn($config);
-        $getJwtProvider = $this->container->get(LcobucciJwtProvider::class)->willReturn($this->tokenProvider->reveal());
+        $this->container->expects($this->exactly(2))->method('get')->willReturnMap([
+            ['config', $config],
+            [LcobucciJwtProvider::class, $this->createMock(TokenProviderInterface::class)],
+        ]);
 
-        $hub = ($this->factory)($this->container->reveal());
+        $hub = ($this->factory)($this->container);
 
         $ref = new ReflectionObject($hub);
         $prop = $ref->getProperty('url');
         $prop->setAccessible(true);
 
         self::assertEquals($expectedHubUrl . '/.well-known/mercure', $prop->getValue($hub));
-        $getConfig->shouldHaveBeenCalledOnce();
-        $getJwtProvider->shouldHaveBeenCalledOnce();
     }
 
     public function provideValidConfigs(): iterable