+++ test/Paginator/PaginatorTest.php
@@ -5,34 +5,33 @@ declare(strict_types=1);
 namespace ShlinkioTest\Shlink\Common\Paginator;
 
 use Pagerfanta\Adapter\AdapterInterface;
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use Prophecy\PhpUnit\ProphecyTrait;
-use Prophecy\Prophecy\ObjectProphecy;
 use Shlinkio\Shlink\Common\Paginator\Paginator;
 
 class PaginatorTest extends TestCase
 {
-    use ProphecyTrait;
-
     private Paginator $paginator;
-    private ObjectProphecy $adapter;
+    private MockObject & AdapterInterface $adapter;
 
     protected function setUp(): void
     {
-        $this->adapter = $this->prophesize(AdapterInterface::class);
-        $this->paginator = new Paginator($this->adapter->reveal());
+        $this->adapter = $this->createMock(AdapterInterface::class);
+        $this->paginator = new Paginator($this->adapter);
     }
 
     /**
+     * @param -1|int<1, max> $maxPage
      * @test
      * @dataProvider providePositiveNumbers
      */
     public function setMaxBehavesAsUsualWhenPositiveNumberIsProvided(int $maxPage): void
     {
+        $this->adapter->expects($this->never())->method('getNbResults');
+
         $this->paginator->setMaxPerPage($maxPage);
 
         self::assertEquals($maxPage, $this->paginator->getMaxPerPage());
-        $this->adapter->getNbResults()->shouldNotHaveBeenCalled();
     }
 
     public function providePositiveNumbers(): iterable
@@ -41,18 +40,18 @@ class PaginatorTest extends TestCase
     }
 
     /**
+     * @param -1|int<1, max> $maxPage
      * @test
      * @dataProvider provideNonPositiveNumbers
      */
     public function setMaxFallsBackToAdapterWhenNonPositiveNumberIsProvided(int $maxPage): void
     {
         $expected = 35;
-        $getFromAdapter = $this->adapter->getNbResults()->willReturn($expected);
+        $this->adapter->expects($this->once())->method('getNbResults')->willReturn($expected);
 
         $this->paginator->setMaxPerPage($maxPage);
 
         self::assertEquals($expected, $this->paginator->getMaxPerPage());
-        $getFromAdapter->shouldHaveBeenCalledOnce();
     }
 
     public function provideNonPositiveNumbers(): iterable
@@ -66,12 +65,11 @@ class PaginatorTest extends TestCase
      */
     public function getMaxReturnsOneWhenAdapterReturnsEmpty(int $adapterNbResults): void
     {
-        $getFromAdapter = $this->adapter->getNbResults()->willReturn($adapterNbResults);
+        $this->adapter->expects($this->once())->method('getNbResults')->willReturn($adapterNbResults);
 
         $this->paginator->setMaxPerPage(-1);
 
         self::assertEquals(1, $this->paginator->getMaxPerPage());
-        $getFromAdapter->shouldHaveBeenCalledOnce();
     }
 
     public function provideEmptyAdapterResults(): iterable