+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 
 ### Changed
 * Migrated infection config to json5
+* Migrated from prophecy to PHPUnit mocks
 
 ### Deprecated
 * *Nothing*